+++ demo-app/app/src/main/java/com/synapseslab/bluegpssdkdemo/DemoApplication.kt
@@ -37,7 +37,9 @@ class DemoApplication : Application() {
             sdkEnvironment = Environment.sdkEnvironment,
             context = this,
             enabledNetworkLogs = BuildConfig.DEBUG,
-            keyCloakParameters = Environment.keyCloakParameters
+
+            // enable this for Keycloak authentication
+            //keyCloakParameters = Environment.keyCloakParameters
         )
     }